refactor: split api key settings transport
